+++ benches/distances/kattis_shortestpath1.rs
@@ -9,7 +9,7 @@ use {
         },
         algo::{
             bellman_ford_moore,
-            dijkstra_dist::Dijkstra,
+            dijkstra_dist::DijkstraDist,
             floyd_warshall,
         },
     },
@@ -33,7 +33,7 @@ fn dijkstra(bencher: Bencher<'_, '_>) {
     let digraph = kattis_shortestpath1_usize();
 
     bencher.bench_local(|| {
-        let mut dijkstra = Dijkstra::new(&digraph, &[0]);
+        let mut dijkstra = DijkstraDist::new(&digraph, &[0]);
         let _ = dijkstra.distances();
     });
 }