+++ src/lib.rs
@@ -118,26 +118,26 @@
 //! A breadth-first search explores the vertices of an unweighted digraph in
 //! order of their distance from a source.
 //!
-//! - [`bfs::Bfs`](algo::bfs::Bfs) iterates over the vertices.
-//! - [`bfs_dist::Bfs`](algo::bfs_dist::Bfs) iterates over the vertices and
-//!   their distance from the source.
-//! - [`bfs_pred::Bfs`](algo::bfs_pred::Bfs) iterates over the vertices and
-//!   their predecessors.
-//! - [`bfs_dist::Bfs::distances`](algo::bfs_dist::Bfs::distances) finds the
+//! - [`Bfs`](algo::bfs::Bfs) iterates over the vertices.
+//! - [`BfsDist`](algo::bfs_dist::BfsDist) iterates over the vertices and their
+//!   distance from the source.
+//! - [`BfsPred`](algo::bfs_pred::BfsPred) iterates over the vertices and their
+//!   predecessors.
+//! - [`BfsDist::distances`](algo::bfs_dist::BfsDist::distances) finds the
 //!   shortest distances.
-//! - [`bfs_pred::Bfs::predecessors`](algo::bfs_pred::Bfs::predecessors) finds
+//! - [`BfsPred::predecessors`](algo::bfs_pred::BfsPred::predecessors) finds
 //!   the predecessors.
-//! - [`bfs_pred::Bfs::shortest_path`](algo::bfs_pred::Bfs::shortest_path)
-//!   finds the shortest path.
+//! - [`BfsPred::shortest_path`](algo::bfs_pred::BfsPred::shortest_path) finds
+//!   the shortest path.
 
 //! ## Depth-First Search (DFS)
 //!
 //! A depth-first search explores the vertices of an unweighted digraph in
 //! order of their distance from a source.
 //!
-//! - [`dfs::Dfs`](algo::dfs::Dfs) iterates over the vertices.
-//! - [`dfs_dist::Dfs`](algo::dfs_dist::Dfs) iterates over the vertices and
-//!   their distance from the source.
+//! - [`Dfs`](algo::dfs::Dfs) iterates over the vertices.
+//! - [`DfsDist`](algo::dfs_dist::DfsDist) iterates over the vertices and their
+//!   distance from the source.
 //!
 //! ## Dijkstra
 //!
@@ -146,13 +146,14 @@
 //!
 //! - [`dijkstra::Dijkstra`](algo::dijkstra::Dijkstra) iterates over the
 //!   vertices.
-//! - [`dijkstra_dist::Dijkstra`](algo::dijkstra_dist::Dijkstra) iterates over
-//!   the vertices and their distance from the source.
-//! - [`dijkstra_pred::Dijkstra`](algo::dijkstra_pred::Dijkstra) iterates over
-//!   the vertices and their predecessors.
-//! - [`dijkstra_dist::Dijkstra::distances`](algo::dijkstra_dist::Dijkstra::distances) finds the shortest distances.
-//! - [`dijkstra_pred::Dijkstra::predecessors`](algo::dijkstra_pred::Dijkstra::predecessors) finds the predecessors.
-//! - [`dijkstra_pred::Dijkstra::shortest_path`](algo::dijkstra_pred::Dijkstra::shortest_path) finds the shortest path.
+//! - [`DijkstraDist`](algo::dijkstra_dist::DijkstraDist) iterates over the
+//!   vertices and their distance from the source.
+//! - [`DijkstraPred`](algo::dijkstra_pred::DijkstraPred) iterates over the
+//!   vertices and their predecessors.
+//! - [`DijkstraDist::distances`](algo::dijkstra_dist::DijkstraDist::distances)
+//!   finds the shortest distances.
+//! - [`DijkstraPred::predecessors`](algo::dijkstra_pred::DijkstraPred::predecessors) finds the predecessors.
+//! - [`DijkstraPred::shortest_path`](algo::dijkstra_pred::DijkstraPred::shortest_path) finds the shortest path.
 //!
 //! ## Floyd-Warshall
 //!
@@ -225,6 +226,7 @@
     clippy::multiple_inherent_impl,
     clippy::panic_in_result_fn,
     clippy::redundant_type_annotations,
+    clippy::renamed_function_params,
     clippy::rest_pat_in_fully_bound_structs,
     clippy::self_named_module_files,
     clippy::unnecessary_self_imports,